redoing check if csv has headers

+++ kv_timer_verifier.py
@@ -143,23 +143,28 @@ def get_all_gpu_memcpy_correlations(json_filename, get_cpu_time=False, est_bandw
         total_storing_bytes /= 1000000000.0 # B --> GB
     #     est_loading_bandwidth = total_loading_bytes / total_loading_cache_time_gpu # GB / s
     #     est_storing_bandwidth = total_storing_bytes / total_storing_cache_time_gpu # GB / s
-    csv_filename = json_filename[:-41]  + '.csv'
-    add_headers = not Path(csv_filename).exists()
-    print(f"csv_filename: {csv_filename}")
-    kv_gpu_percent = int(json_filename.split('-')[9])
-    with open(csv_filename, 'w', newline='') as csvfile:
-            fieldnames = ['kv_gpu_percent', 'tot_loading_time_gpu (s)', 'tot_storing_time_gpu (s)', 'tot_loading_time_cpu (s)','tot_storing_time_cpu (s)',  'total_loading_bytes (GB)', 'total_storing_bytes (GB)']
-            writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
-            # if add_headers:
-            writer.writeheader()
 
-            writer.writerow({'kv_gpu_percent': kv_gpu_percent, 
+    # Check if file exists to write header only once
+    csv_filename = json_filename[:-41]  + '.csv'
+    kv_gpu_percent = int(json_filename.split('-')[9])
+    fieldnames = ['kv_gpu_percent', 'tot_loading_time_gpu (s)', 'tot_storing_time_gpu (s)', 'tot_loading_time_cpu (s)','tot_storing_time_cpu (s)',  'total_loading_bytes (GB)', 'total_storing_bytes (GB)']
+
+    if not os.path.exists(csv_filename):
+        with open(csv_filename, 'w', newline='') as csvfile:
+            writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+            writer.writeheader()
+    
+    # Open the file in append mode ('a')
+    with open(filename, 'a', newline='') as csvfile:
+        writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+        writer.writerow({'kv_gpu_percent': kv_gpu_percent, 
                 'tot_loading_time_gpu (s)': total_loading_cache_time_gpu, 
                 'tot_storing_time_gpu (s)': total_storing_cache_time_gpu, 
                 'tot_loading_time_cpu (s)': total_loading_cache_time_cpu, 
                 'tot_storing_time_cpu (s)': total_storing_cache_time_cpu, 
                 'total_loading_bytes (GB)': total_loading_bytes, 
                 'total_storing_bytes (GB)': total_storing_bytes})
+
 
     return total_loading_cache_time_gpu, total_storing_cache_time_gpu, total_loading_cache_time_cpu, total_storing_cache_time_cpu, total_loading_bytes, total_storing_bytes # in s
